Extract utils functions for tests
