+++ embed-code-go/test/embedding_test/embedding_test.go
@@ -21,82 +21,14 @@ package embedding_test
 import (
 	"embed-code/embed-code-go/configuration"
 	"embed-code/embed-code-go/embedding"
+	"embed-code/embed-code-go/test/utils"
 	"fmt"
-	"io"
 	"os"
 	"path/filepath"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
 )
-
-// Removes directory and all it's subdirectories if exists, does nothing if not exists.
-func cleanupDir(dir string) {
-	if _, err := os.Stat(dir); err == nil {
-		err = os.RemoveAll(dir)
-		if err != nil {
-			panic(err)
-		}
-	}
-}
-
-// Copies directory from source path to target path with all subdirs and children.
-func copyDirRecursive(source string, target string) {
-	info, err := os.Stat(source)
-	if err != nil {
-		panic(err)
-	}
-
-	err = os.MkdirAll(target, info.Mode())
-	if err != nil {
-		panic(err)
-	}
-
-	entries, err := os.ReadDir(source)
-	if err != nil {
-		panic(err)
-	}
-
-	for _, entry := range entries {
-		sourcePath := filepath.Join(source, entry.Name())
-		targetPath := filepath.Join(target, entry.Name())
-
-		if entry.IsDir() {
-			copyDirRecursive(sourcePath, targetPath)
-		} else {
-			err = copyFile(sourcePath, targetPath)
-			if err != nil {
-				panic(err)
-			}
-		}
-	}
-}
-
-func copyFile(source string, target string) (err error) {
-	sourceFile, err := os.Open(source)
-	if err != nil {
-		return
-	}
-	defer sourceFile.Close()
-
-	targetFile, err := os.Create(target)
-	if err != nil {
-		return
-	}
-	defer func() {
-		cerr := targetFile.Close()
-		if err == nil {
-			err = cerr
-		}
-	}()
-
-	if _, err = io.Copy(targetFile, sourceFile); err != nil {
-		return
-	}
-
-	err = os.Chmod(target, os.FileMode(0666))
-	return
-}
 
 type EmbeddingInstructionTestsPreparator struct {
 	rootDir  string
@@ -121,12 +53,12 @@ func newEmbeddingInstructionTestsPreparator() EmbeddingInstructionTestsPreparato
 func (testPreparator EmbeddingInstructionTestsPreparator) setup() {
 	config := buildConfigWithPreparedFragments()
 	os.Chdir(testPreparator.rootDir)
-	copyDirRecursive("./test/resources/docs", config.DocumentationRoot)
+	utils.CopyDirRecursive("./test/resources/docs", config.DocumentationRoot)
 }
 
 func (testPreparator EmbeddingInstructionTestsPreparator) cleanup() {
 	config := buildConfigWithPreparedFragments()
-	cleanupDir(config.DocumentationRoot)
+	utils.CleanupDir(config.DocumentationRoot)
 	os.Chdir(testPreparator.testsDir)
 }
